Changed function def and comment for new behaviour.
Now the db_execute function doesn't require the database argument.
The db_helper file saves the database that it opens for future use.
This means currently we only support opening and using a Single database.

Also made the ignore topics match for both an ends with and a starts with match.
So you can specify ignore things starting with X or where the topic is X plus any root.
It does Both these checks, so make sure you don't have a rule that matches two topics by starts with one one and ends with on another.
Use `root+"stuff"` where stuff is the topic that the message ends with to match.
Otherwise just use `"stuff"` where stuff is the topic the message starts with to match.