Added temperature module
Listens on "temperatures/" root for all topics.
Records the topic as the host name and the payload as a float temperature.
So "temperatures/temp4 = 10.2" means the SQL database gets "datetime | temp4 | location | 10.2"
Where datetime is the time of message receive, and location is a dictionary linked to the host names
inside the temperature.py file.